+++ encoderfile/templates/main.rs.tera
@@ -1,11 +1,6 @@
 // Generated by encoderfile ❤️ DO NOT MODIFY
 
-use encoderfile_core::{
-    factory,
-    cli_entrypoint,
-};
-
-factory! {
+encoderfile_core::factory! {
     "{{ model_weights_path }}",
     "{{ tokenizer_path }}",
     "{{ model_config_path }}",
@@ -14,15 +9,4 @@ factory! {
     {% if transform %}Some(r###"
 {{ transform }}
     "###){% else %}None{% endif %},
-}
-
-fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    cli_entrypoint(
-        &assets::MODEL_WEIGHTS,
-        assets::MODEL_CONFIG_JSON,
-        assets::TOKENIZER_JSON,
-        assets::MODEL_TYPE_STR,
-        assets::MODEL_ID,
-        assets::TRANSFORM,
-    )
 }
